tech-debt(web): consolidate date/contribution/focus-trap/og-script helpers
Pull together four sets of duplicated web-only helpers so each pattern
has a single source of truth.

- New `apps/web/lib/utils/dates.ts` exports `formatDateMonDay` (admin
  chart axes) and `formatDateKey` (local-time YYYY-MM-DD). Replaces 6
  inline copies in admin charts and 5 inline copies in
  share-assets / contribution / recap utilities.
- New `fillContributionDays` export on `apps/web/lib/utils/recap.ts`
  replaces three byte-identical copies in the recap page, RecapCard,
  and recap-image renderer.
- New `apps/web/components/app/shared/useFocusTrap.ts` hook replaces
  three nearly-identical inline focus traps in SubmitPromptWidget,
  ImageLightbox, and SuggestCompanyWidget. The shared selector
  matches the broader form-aware variant; the lightbox has no
  textareas/inputs so the wider selector is a no-op there.
- New `apps/web/scripts/_lib.ts` exports `ensureCleanOutputDir(dir)`
  used by all three OG generator scripts.

No behavior change. tsc + apps/web vitest pass.
