+++ apps/web/app/admin/components/UserSignupsChart.tsx
@@ -11,6 +11,7 @@ import {
   ResponsiveContainer,
 } from "recharts";
 import { useAdminTheme } from "./AdminShell";
+import { formatDateMonDay as formatDate } from "@/lib/utils/dates";
 
 interface SignupRow {
   date: string;
@@ -22,11 +23,6 @@ const RANGES = [
   { key: "30d", label: "30D", days: 30 },
   { key: "all", label: "All", days: Infinity },
 ] as const;
-
-function formatDate(dateStr: string) {
-  const d = new Date(dateStr + "T00:00:00");
-  return d.toLocaleDateString("en-US", { month: "short", day: "numeric" });
-}
 
 function Skeleton() {
   return (
