+++ apps/web/components/app/prompts/SubmitPromptWidget.tsx
@@ -6,6 +6,7 @@ import Link from "next/link";
 import { X, Sparkles } from "lucide-react";
 import { Button } from "@/components/ui/Button";
 import { Textarea } from "@/components/ui/Textarea";
+import { useFocusTrap } from "@/components/app/shared/useFocusTrap";
 import { timeAgo } from "@/lib/utils/format";
 
 const MAX_PROMPT_LENGTH = 2000;
@@ -71,27 +72,7 @@ export function SubmitPromptWidget({ username }: SubmitPromptWidgetProps) {
     }
   }, [open]);
 
-  useEffect(() => {
-    if (!open) return;
-    function onFocusTrap(e: KeyboardEvent) {
-      if (e.key !== "Tab" || !dialogRef.current) return;
-      const focusable = dialogRef.current.querySelectorAll<HTMLElement>(
-        'button, textarea, input, [href], [tabindex]:not([tabindex="-1"])',
-      );
-      if (focusable.length === 0) return;
-      const first = focusable[0];
-      const last = focusable[focusable.length - 1];
-      if (e.shiftKey && document.activeElement === first) {
-        e.preventDefault();
-        last.focus();
-      } else if (!e.shiftKey && document.activeElement === last) {
-        e.preventDefault();
-        first.focus();
-      }
-    }
-    document.addEventListener("keydown", onFocusTrap);
-    return () => document.removeEventListener("keydown", onFocusTrap);
-  }, [open]);
+  useFocusTrap(dialogRef, open);
 
   useEffect(() => {
     if (!flash) return;
