+++ apps/web/lib/share-assets/profile-card-data.ts
@@ -1,5 +1,6 @@
 import type { SupabaseClient } from "@supabase/supabase-js";
 import { prettifyModel } from "@/lib/utils/post-share";
+import { formatDateKey } from "@/lib/utils/dates";
 
 export interface ProfileShareCardData {
   username: string;
@@ -18,13 +19,6 @@ interface ProfileRow {
   username: string | null;
   display_name: string | null;
   is_public: boolean;
-}
-
-function formatDate(date: Date) {
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, "0");
-  const d = String(date.getDate()).padStart(2, "0");
-  return `${y}-${m}-${d}`;
 }
 
 function resolvePrimaryModel(
@@ -64,7 +58,7 @@ export async function getProfileShareCardData(
         .from("daily_usage")
         .select("date, cost_usd")
         .eq("user_id", profile.id)
-        .gte("date", formatDate(heatmapStart))
+        .gte("date", formatDateKey(heatmapStart))
         .order("date", { ascending: true }),
       supabase
         .from("daily_usage")
@@ -74,7 +68,7 @@ export async function getProfileShareCardData(
         .from("daily_usage")
         .select("date, output_tokens, models")
         .eq("user_id", profile.id)
-        .gte("date", formatDate(recentStart)),
+        .gte("date", formatDateKey(recentStart)),
       supabase.rpc("calculate_user_streak", { p_user_id: profile.id }),
     ]);
