+++ apps/web/lib/utils/recap-image.tsx
@@ -1,45 +1,10 @@
-import type { RecapData } from "./recap";
+import { fillContributionDays, type RecapData } from "./recap";
 import { formatCurrency, formatTokens, getCellColor } from "./format";
 import {
   LIGHT_PALETTE,
   DARK_PALETTE,
   type RecapPalette,
 } from "@/lib/recap-backgrounds";
-
-/** Fill in missing days with $0 entries — only up to today (no future days) */
-function fillContributionDays(
-  data: { date: string; cost_usd: number }[],
-  totalDays: number,
-  period: "week" | "month"
-): { date: string; cost_usd: number }[] {
-  const lookup = new Map(data.map((d) => [d.date, d.cost_usd]));
-  const now = new Date();
-  let start: Date;
-
-  if (period === "week") {
-    const day = now.getDay();
-    const mondayOffset = day === 0 ? -6 : 1 - day;
-    start = new Date(now);
-    start.setDate(now.getDate() + mondayOffset);
-  } else {
-    start = new Date(now.getFullYear(), now.getMonth(), 1);
-  }
-
-  // Cap at today — don't render future days
-  const msPerDay = 86400000;
-  const daysSinceStart =
-    Math.floor((now.getTime() - start.getTime()) / msPerDay) + 1;
-  const cappedDays = Math.min(totalDays, daysSinceStart);
-
-  const result: { date: string; cost_usd: number }[] = [];
-  for (let i = 0; i < cappedDays; i++) {
-    const d = new Date(start);
-    d.setDate(start.getDate() + i);
-    const key = `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, "0")}-${String(d.getDate()).padStart(2, "0")}`;
-    result.push({ date: key, cost_usd: lookup.get(key) ?? 0 });
-  }
-  return result;
-}
 
 /**
  * Renders the recap card JSX for use in next/og ImageResponse.
