+++ apps/web/lib/utils/recap.ts
@@ -1,4 +1,5 @@
 import type { SupabaseClient } from "@supabase/supabase-js";
+import { formatDateKey } from "@/lib/utils/dates";
 
 export interface RecapData {
   total_cost: number;
@@ -59,8 +60,8 @@ function getPeriodRange(period: "week" | "month"): {
         : `, ${now.getFullYear()}`;
 
     return {
-      start: formatDate(monday),
-      end: formatDate(sunday),
+      start: formatDateKey(monday),
+      end: formatDateKey(sunday),
       totalDays: 7,
       label: `My Week in Claude Code · ${fmt(monday)}–${fmt(sunday)}${yearSuffix}`,
     };
@@ -76,18 +77,50 @@ function getPeriodRange(period: "week" | "month"): {
   });
 
   return {
-    start: formatDate(firstDay),
-    end: formatDate(lastDay),
+    start: formatDateKey(firstDay),
+    end: formatDateKey(lastDay),
     totalDays: lastDay.getDate(),
     label: `My Month in Claude Code · ${monthName}`,
   };
 }
 
-function formatDate(d: Date): string {
-  const y = d.getFullYear();
-  const m = String(d.getMonth() + 1).padStart(2, "0");
-  const day = String(d.getDate()).padStart(2, "0");
-  return `${y}-${m}-${day}`;
+/**
+ * Fill in missing days with $0 entries — only up to today (no future days).
+ * Used by the recap card components and the OG image renderer to produce a
+ * complete contribution strip for the current week or month.
+ */
+export function fillContributionDays(
+  data: { date: string; cost_usd: number }[],
+  totalDays: number,
+  period: "week" | "month",
+): { date: string; cost_usd: number }[] {
+  const lookup = new Map(data.map((d) => [d.date, d.cost_usd]));
+  const now = new Date();
+  let start: Date;
+
+  if (period === "week") {
+    const day = now.getDay();
+    const mondayOffset = day === 0 ? -6 : 1 - day;
+    start = new Date(now);
+    start.setDate(now.getDate() + mondayOffset);
+  } else {
+    start = new Date(now.getFullYear(), now.getMonth(), 1);
+  }
+
+  // Cap at today — don't render future days
+  const msPerDay = 86400000;
+  const daysSinceStart =
+    Math.floor((now.getTime() - start.getTime()) / msPerDay) + 1;
+  const cappedDays = Math.min(totalDays, daysSinceStart);
+
+  const result: { date: string; cost_usd: number }[] = [];
+  for (let i = 0; i < cappedDays; i++) {
+    const d = new Date(start);
+    d.setDate(start.getDate() + i);
+    const key = `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, "0")}-${String(d.getDate()).padStart(2, "0")}`;
+    result.push({ date: key, cost_usd: lookup.get(key) ?? 0 });
+  }
+  return result;
 }
 
 export async function getRecapData(
